% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{get_geometry}
\alias{get_geometry}
\title{Get base geometries}
\usage{
get_geometry(obj)
}
\arguments{
\item{obj}{tissot object}
}
\value{
Geometry as objects of class sf
}
\description{
This function returns the geometries, from which Tissot indicatrix
was generated in the first place.
}
