% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{print}
\alias{print}
\alias{print.tissot}
\title{This function prints Tissot indicatrix circles that were generated}
\usage{
\method{print}{tissot}(x, ...)
}
\arguments{
\item{x}{tissot object}

\item{...}{ignored}
}
\description{
This function prints Tissot indicatrix circles that were generated
}
