% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{summary}
\alias{summary}
\alias{summary.tissot}
\title{Summarize indicatrix}
\usage{
\method{summary}{tissot}(object, ...)
}
\arguments{
\item{object}{tissot object}

\item{...}{ignored}
}
\description{
This function summarizes Tissot indicatrix circles details
}
