% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{tissot}
\alias{tissot}
\title{This function generates Tissot's indicatrix}
\usage{
tissot(geom, circles_den = c(-1), circle_size = "auto")
}
\arguments{
\item{geom}{input geometry for which tissot should be generated}

\item{circles_den}{Specifies the number of indicatrix circles that are generated.
Default is c(-1) which corresponds to automatic density.
Manually, number of circles in x and y direction can be
specified through an array, e.g. c(20, 10).}

\item{circle_size}{Specifies the size of each indicatrix circle. Default it is
automatic. Manually, size in Meters should be specified.}
}
\value{
circles as objects of class sf
}
\description{
This function returns equally-sized circles (radius = 300km) as sf-polygons
to use as Tissot's indicatrix. The circles are located on a regular grid in
lon/lat every 20 degrees.
}
