% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissot.R
\name{get_geometry}
\alias{get_geometry}
\title{This function returns the original geometries, from which Tissot indicatrix
was generated}
\usage{
get_geometry(obj)
}
\arguments{
\item{obj}{tissot object}
}
\value{
Original geometry as objects of class sf
}
\description{
This function returns the original geometries, from which Tissot indicatrix
was generated
}
